Codigo/service_core_projects/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
django-cors-headers>=4.3.0
djangorestframework
django-cors-headers